14329 Fix issues from PR comments
